# swagger.yaml
openapi: 3.0.0
info:
  title: EasyTrip
  version: 1.0.0
  description: EasyTrip official API
  license:
    name: Licensed Under MIT
    url: https://spdx.org/licenses/MIT.html
  contact:
    name: JSONPlaceholder
    url: https://jsonplaceholder.typicode.com
servers:
  - url: http://localhost:12349
    description: EasyTrip Facade server
  - url: http://localhost:12348
    description: Path Search server
  - url: http://localhost:12347
    description: Point Search server
  - url: http://localhost:12346
    description: OSM Tools server
  - url: http://localhost:12345
    description: Auth Tools server
paths:
...
  '/v2/combined/about':
    get:
      summary: Get information about weather, bike sharing and points of interest in a certain location based on some choices
      description: Return the information about weather, bike sharing and points of interest in a certain location (can choose if all three fields will be returned or if some of them) based on some interests
      tags:
        - Point Search
      parameters:
        - $ref: '#/components/parameters/address'
        - $ref: '#/components/parameters/weather'
        - $ref: '#/components/parameters/bikes'
        - $ref: '#/components/parameters/interest'
      responses:
        '200':
          description: Object about requested information and points of interest
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    $ref: '#/components/schemas/start'
                  weather:
                    $ref: '#/components/schemas/weather'
                  bike:
                    $ref: '#/components/schemas/bikes'
                  poi:
                    $ref: '#/components/schemas/poiArray'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
...
    address:
      description: Address to search
      name: address
      in: query
      schema:
        type: string
      required: true
      example: "Verona"
...
    weather:
      name: weather
      in: query
      description: describe if in the request user wants weather forecasts
      schema:
        $ref: '#/components/schemas/weatherIndicator'
      example: false
    bikes:
      name: bikes
      in: query
      description: describe if in the request user wants bike sharing
      schema:
        $ref: '#/components/schemas/bikesIndicator'
      example: false
    interest:
      name: interest
      in: query
      description: User interests. If not specified it the searchs won't be performed
      schema:
        $ref: '#/components/schemas/interests'
...
  schemas:
...
    weatherIndicator:
      type: boolean
      description: describe if in the request user wants weather forecasts
      default: false
      example: false
    bikesIndicator:
      type: boolean
      description:  describe if in the request user wants bike sharing
      default: false
      example: false
    interests:
      type: array
      items:
        type: string
        description: interest of the search for the points of interests
        enum:
          - "sustenance"
          - "education"
          - "entertainment"
          - "tourism"
          - "accomodation"
...
    geocode:
      type: object
      description: Geocode result
      properties:
        point:
          type: object
          properties:
            lat:
              $ref: '#/components/schemas/latitude'
            lng:
              $ref: '#/components/schemas/longitude'
          description: Coordinates
        osm_id:
          type: string
          description: The OSM ID of the entity
        osm_type:
          type: string
          description: N = node, R = relation, W = way
        osm_key:
          type: string
          description: The OSM key of the entity
        osm_value:
          type: string
          description: The OSM value of the entity
        name:
          type: string
          description: The name of the entity. Can be a boundary, POI, address, etc
        country:
          type: string
          description: The country of the address
        countrycode:
          type: string
          description: Code representation of country
        city:
          type: string
          description: The city of the address
        state:
          type: string
          description: The state of the address
        street:
          type: string
          description: The street of the address
        housenumber:
          type: string
          description: The housenumber of the address
        postcode:
          type: string
          description: The postcode of the address
    start:
      allOf:
        - $ref: '#/components/schemas/geocode'
        - example:
            $ref: 'docExamples/arenaGeocode.json'
...
    bikes:
      type: object
      properties:
        company:
          type: string
          example: 'Comunicare S.r.l.'
          description: The company providing the bike sharing service
        gbfs_href:
          type: string
          description: Website url of the bike sharing service
        href:
          type: string
          example: '/v2/networks/e-motion-trento'
          description: resource URI on bike sharing API Website
        id:
          type: string
          example: 'e-motion-trento'
          description: bike sharing identifier
        location:
          type: object
          properties:
            city:
              type: string
              example: 'Trento'
              description: City where the bike sharing service is available
            country:
              type: string
              example: 'IT'
              description: Country where the city is located
            latitude:
              type: number
              example: 46.06643205823519
              description: latitute coordinate value
            longitude:
              type: number
              example: 11.122145390351879
              description: longitude coordinate value
        name:
          type: string
          example: 'e.motion'
          description: name of the bike sharing service
        source:
          type: string
          example: 'https://www.bicincitta.com/frmLeStazioni.aspx?ID=187'
          description: link web of the bike sharing provider
    weather:
      type: object
      properties:
        current:
          type: string
          example: 'Clear sky'
          description: Current weather status of address
        forecasts:
          type: array
          example: 
            - "Few clouds"
            - "Clear Sky"
            - "Mix snow/rain"
            - "Flurries"
            - "Flurries"
            - "Few clouds"
            - "Broken clouds"
            - "Flurries"
            - "Few clouds"
            - "Broken clouds"
            - "Clear Sky"
            - "Clear Sky"
            - "Light snow"
            - "Heavy snow"
            - "Heavy snow"
            - "Mix snow/rain"
          description: Weather forecasts of the next 16 days
          items:
            type: string
    latitude:
      type: number
      description: Latitude
      minimum: -90
      maximum: 90
    longitude:
      type: number
      description: Longitude
      minimum: -180
      maximum: 180
...
    poiArray:
      type: array
      description: Array of POI objects
      items:
        $ref: '#/components/schemas/poi'
      example:
        - $ref: 'docExamples/poi.json'
...
    poi:
      type: object
      properties:
        type:
          type: string
          enum:
            - node
        id:
          type: integer
        lat:
          $ref: '#/components/schemas/latitude'
        lon:
          $ref: '#/components/schemas/longitude'
        tags:
          type: object
          description: Open Street Map tags
...
  responses:
...
    BadRequest:
      description: Bad request
...